requirements.txt
aniso8601==8.0.0
cffi==1.14.0
click==7.1.2
Flask==1.1.2
Flask-Ext==0.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
greenlet==0.4.16
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pycparser==2.20
pymongo==3.10.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.1
zope.event==4.4
zope.interface==5.1.0
